requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
dateparser==1.2.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pip==24.2
propcache==0.2.1
pycryptodome==3.21.0
python-binance==1.0.25
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
schedule==1.2.2
setuptools==75.1.0
six==1.17.0
tzlocal==5.2
urllib3==2.3.0
websockets==14.1
wheel==0.44.0
yarl==1.18.3
